Renamed EventDetailsOverview to EventDetailsViewModel (+ktfmt format)
