+++ app/src/main/java/com/github/se/eventradar/ui/home/Home.kt
@@ -1,6 +1,5 @@
 package com.github.se.eventradar.ui.home
 
-import android.annotation.SuppressLint
 import android.widget.Toast
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -54,7 +53,6 @@ import com.github.se.eventradar.ui.BottomNavigationMenu
 import com.github.se.eventradar.ui.navigation.NavigationActions
 import com.github.se.eventradar.ui.navigation.TOP_LEVEL_DESTINATIONS
 import java.time.LocalDateTime
-
 
 @Composable
 fun HomeScreen() { // viewModel: EventViewModel = viewModel() + Navigation
